Update thetvdb-java [1.3.0 -> 1.4.2].
- Fixes crash due to episode DVD number type not being Double.
- lastUpdated is always Long.
- retrofit 2.2.0.

+++ SeriesGuide/src/main/java/com/battlelancer/seriesguide/thetvdbapi/TvdbTools.java
@@ -686,7 +686,7 @@ public class TvdbTools {
                      * a month (ensures show air time changes get stored).
                      */
                     Long lastEditEpoch = localEpisodeIds.get(episodeId);
-                    Integer lastEditEpochNew = episode.lastUpdated;
+                    Long lastEditEpochNew = episode.lastUpdated;
                     if (lastEditEpoch != null && lastEditEpochNew != null
                             && (lastEditEpoch < lastEditEpochNew
                             || dateLastMonthEpoch < lastEditEpoch)) {
